Crate sator_stake_viewer

Stake for viewers:

user stakes has minimal amounts to stake and time to stake to fit specified rank, which is associate with multiplier

user can stake any amount, so minimal time should be not less than minimal for smallest rank

amount can be less than minimal rank

adding amount resets the timer

total 4 ranks

stake_authority is derived operations are signed by on chain stake derived signature

stake_pool’s token_account is derived from and owned by stake_authority

stake_account for each user is derived

 let stake_pool = Pubkey::new_unique();
 let (stake_authority, _) = Pubkey::find_program_address(&[&stake_pool.to_bytes()[..32]], &stake_viewer_program_id());
 let token_account_stake_target = Pubkey::create_with_seed(
     &stake_authority,
     "ViewerStakePool::token_account",
     &spl_token::id(),
 );
 // Pubkey.to_string is longer than 32 chars limit in Solana for seed
 // ETH compatible something
 let seed = user_wallet.to_bytes();
 let seed = bs58::encode(&seed[..20]).into_string();
 let stake_account = Pubkey::create_with_seed(stake_authority, &seed, &stake_viewer_program_id());

Modules

errors

instruction

Program instruction state

state

Program owned state

types

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

id

Returns the program ID

stake_viewer_program_id