Function sator_stake_viewer::instruction::lock

pub fn lock(
    wallet: &Pubkey, 
    stake: &Pubkey, 
    token_account_source: &Pubkey, 
    input: LockInput
) -> Result<(Instruction, Pubkey), ProgramError>

Creates Instruction::Lock instruction which transfer amount from token_account_source to token_account_stake_target. If lock_account initialized, resets timer.

Accounts:

system_program - program, implicit

sysvar_rent - program, implicit to create lock_account which will be rent except if needed

clock - program, implicit

spl_token - program, implicit

wallet - signer, payer

stake -

stake_authority - derived as in Instruction::InitializeStake

token_account_source - mutable

token_account_stake_target - derived, mutable, implicit

lock_account - implicit, derived, mutable from wallet and stake_authority

Notes:

current design does not creates token account to lock tokens, just counts amount in lock.

lock instruction is same instruction as initialize lock, so it could be made different by having separate lock (it will reduce amount of accounts during lock invocation)